+++ scripts/breaking_changes_checker/breaking_changes_allowlist.py
@@ -6,7 +6,7 @@
 # --------------------------------------------------------------------------------------------
 
 
-RUN_BREAKING_CHANGES_PACKAGES = []
+RUN_BREAKING_CHANGES_PACKAGES = ["azure-mgmt-*"]
 
 
 # See Readme for ignore format
@@ -20,6 +20,8 @@ IGNORE_BREAKING_CHANGES = {
         ("ChangedParameterKind", "*", "*", "*", "filter"),
         ("ChangedParameterKind", "*", "*", "*", "skip"),
         ("RemovedOrRenamedPositionalParam", "*", "*", "*", "maxpagesize"),
+        # msrest model bases on vendored _serialization.model instead of msrest.Model which no longer has "validate"
+        ("RemovedOrRenamedClassMethod", "*", "*", "validate"),
         # Changes due to not using msrest anymore
         ("RemovedOrRenamedClassMethod", "*", "*", "as_dict"),
         ("RemovedOrRenamedClassMethod", "*", "*", "deserialize"),